Main Homepage UI Including Problems
I remade the main page of the website, the only difference (that im aware of) being "Buy in Bulk" rather than "Qty Discount". I tried to keep the same problems so we can go step by step and figure out what we want to change. This will also work as a navigation/hub to other pages that look really bad like any of their product pages.